refactor: separate Hermes adapter lifecycle
