+++ tests/adapters/test_hermes_config_builder.py
@@ -17,7 +17,7 @@ if sys.version_info >= (3, 14):
         allow_module_level=True,
     )
 
-from nemo_fabric_adapters.hermes import adapter
+from nemo_fabric_adapters.hermes import configuration
 
 
 def test_build_hermes_config_omits_unset_values_without_hermes_agent():
@@ -33,7 +33,7 @@ def test_build_hermes_config_omits_unset_values_without_hermes_agent():
         }
     )
 
-    config = adapter.build_hermes_config(config, workspace=".")
+    config = configuration.build_hermes_config(config, workspace=".")
 
     assert config["model"] == {
         "provider": "nvidia",
@@ -61,7 +61,7 @@ def test_write_hermes_config_round_trips_without_pyyaml(
         }
     )
 
-    config_path, config = adapter.write_hermes_config(
+    config_path, config = configuration.write_hermes_config(
         agent_config,
         tmp_path / "hermes-home",
         workspace=".",
